+++ R/create.R
@@ -147,6 +147,9 @@ sf_create_soap <- function(input_data,
     resultset <- bind_rows(resultset, this_set)
   }
   resultset <- resultset %>%
+    select(any_of(c("id", "success", "created", 
+                    "errors.statusCode", "errors.message")), 
+           everything()) %>% 
     type_convert(col_types = cols(.default = col_guess()))
   return(resultset)
 }