+++ R/update.R
@@ -142,6 +142,9 @@ sf_update_soap <- function(input_data,
     resultset <- bind_rows(resultset, this_set)
   }
   resultset <- resultset %>%
+    select(any_of(c("id", "success", "created", 
+                    "errors.statusCode", "errors.message")), 
+           everything()) %>%     
     type_convert(col_types = cols(.default = col_guess()))
   return(resultset)
 }