+++ man-roxygen/include_details.R
@@ -1,0 +1,2 @@
+#' @param include_details \code{logical}; an indicator applying to a synchronous 
+#' indicating whether the run should return summary data with details.