+++ man-roxygen/label.R
@@ -1,0 +1,6 @@
+#' @param label \code{logical}; an indicator of whether the returned data should 
+#' be the label (i.e. formatted value) or the actual value. By default, the labels  
+#' are returned because these are what appear in the Salesforce dashboard and 
+#' more closely align with the column names. For example, "Account.Name" label 
+#' may be \code{"Account B"} and the value \code{0016A0000035mJEQAY}. The former 
+#' (label) more accurately reflects the "Account.Name".